Allow custom icons for categories
The previous way to establish an icon for a user made category was
not correct. It appended a JSON element to a list, unknowing of the
fact that the category icon was a QVariant object.

A minimal, unoptimized correction is proposed in this commit: set again
the category icon via a function that will directly write to the
properties of the object.