Add CategoryButton factory, redefine its behavior
Add a factory for CategoryButton to leverage the amount of code in this
qml file. As a byproduct, remove repeated code.

Redefine behavior on CategoryButton's change category signal - when
searching, it actually should work instead of appearing as disabled.
This makes more sense (why would anyone want them not to work?)